src/main/java/org/example/Solution98.java
package org.example;

/**
 * Validate Binary Search Tree
 */
public class Solution98 {
    public boolean isValidBST(TreeNode root) {
        if (root == null) {
            return true;
        }
        if (!isValidBST(root.left)) {
            return false;
        }
        if (min != null && root.val <= min) {
            return false;
        }
        min = root.val;
        return isValidBST(root.right);
    }

    private Integer min = null;

    public static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;

        TreeNode() {
        }

        TreeNode(int val) {
            this.val = val;
        }

        TreeNode(int val, TreeNode left, TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }
    }
}
